Extract crawl initialization as separate MVC widget
Extract many static methods from controller to QtExtensions
